+++ core/hw/flashrom/flashrom.h
@@ -493,11 +493,11 @@ struct DCFlashChip : MemChip
 		bool valid = true;
 		char sysinfo[16];
 
-		for (int i = 0; i < sizeof(sysinfo); i++)
+		for (size_t i = 0; i < sizeof(sysinfo); i++)
 			sysinfo[i] = Read8(0x1a000 + i);
 		valid = valid && memcmp(&sysinfo[5], "Dreamcast  ", 11) == 0;
 
-		for (int i = 0; i < sizeof(sysinfo); i++)
+		for (size_t i = 0; i < sizeof(sysinfo); i++)
 			sysinfo[i] = Read8(0x1a0a0 + i);
 		valid = valid && memcmp(&sysinfo[5], "Dreamcast  ", 11) == 0;
